Install powerlevel10k zsh theme (#4)
* Use --greedy option when upgrading brew packages

This commit changes the update-mac.sh script to execute `brew upgrade --greedy`
to ensure that we upgrade packages/apps even when they have their own
upgrade mechanisms

* Use vim as the default editor

Current configuration would use neovim as the default editor
but given I have not yet configured it, there is no theme or
configurations applied. Prefer vim for the time being, until neovim is
setup or I decide not to use it at all.

* Delete .p10k.zsh file

powerlevel10k is effectively discontinued and I didn't configure it yet
in these dotfiles. I don't have a need for it, and so deleting any
references.

* Generate powerlevel10k configuration, delete zsh theme

* Add git aliases and default commit template

* Add Terminal theme files

* Add Xcode themes

* Install and configure powerlevel10k

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>